task_11/requirements.txt
cohere
sentence-transformers
transformers
langchain
openai
python-dotenv
numpy
scipy


